# public/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OM Paye API",
        "description": "API pour OM Paye - Système de paiement mobile",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/compte": {
            "get": {
                "summary": "Informations du compte utilisateur",
                "operationId": "95b6694a3fec01268d56823f5fcb7e72",
                "responses": {
                    "200": {
                        "description": "Informations du compte",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "solde": {
                                            "description": "Solde en centimes",
                                            "type": "integer"
                                        },
                                        "type": {
                                            "description": "Type de compte",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/historique": {
            "get": {
                "summary": "Historique des transactions",
                "operationId": "b40f820f0e2f668b39c34be4a271af95",
                "responses": {
                    "200": {
                        "description": "Liste des transactions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "montant": {
                                                        "type": "integer"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/transactions/depot": {
            "post": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Effectuer un dépôt d'argent",
                "operationId": "69adea14cdaed7ebf3b765eb7d733424",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "montant",
                                    "agent_id"
                                ],
                                "properties": {
                                    "montant": {
                                        "description": "Montant en XOF (5000 = 5000 XOF)",
                                        "type": "integer",
                                        "example": 5000
                                    },
                                    "agent_id": {
                                        "description": "ID de l'agent distributeur obligatoire",
                                        "type": "string",
                                        "example": "uuid"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Dépôt réussi"
                    },
                    "400": {
                        "description": "Montant invalide ou agent invalide"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/transactions/retrait": {
            "post": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Effectuer un retrait d'argent",
                "operationId": "5f0d670067719928d88a7dd91657ddd2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "montant",
                                    "agent_id",
                                    "pin"
                                ],
                                "properties": {
                                    "montant": {
                                        "description": "Montant en XOF",
                                        "type": "integer",
                                        "example": 2000
                                    },
                                    "agent_id": {
                                        "type": "string",
                                        "example": "uuid"
                                    },
                                    "pin": {
                                        "type": "string",
                                        "example": "1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Retrait réussi"
                    },
                    "400": {
                        "description": "PIN invalide ou solde insuffisant"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/transactions/transfert": {
            "post": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Transférer de l'argent",
                "operationId": "a4932b3c77b98e547cb00db6e93c802f",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "telephone_dest",
                                    "montant",
                                    "pin"
                                ],
                                "properties": {
                                    "telephone_dest": {
                                        "type": "string",
                                        "example": "781234567"
                                    },
                                    "montant": {
                                        "description": "Montant en XOF",
                                        "type": "integer",
                                        "example": 5000
                                    },
                                    "pin": {
                                        "type": "string",
                                        "example": "1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Solde insuffisant ou PIN incorrect"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/transactions/paiement": {
            "post": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Effectuer un paiement via QR",
                "operationId": "918a290b059547769c9da3789d860cc7",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code_qr",
                                    "montant",
                                    "pin"
                                ],
                                "properties": {
                                    "code_qr": {
                                        "type": "string",
                                        "example": "OM-ABC123"
                                    },
                                    "montant": {
                                        "description": "Montant en XOF",
                                        "type": "integer",
                                        "example": 10000
                                    },
                                    "pin": {
                                        "type": "string",
                                        "example": "1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Paiement réussi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "QR invalide ou solde insuffisant"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token JWT pour l'authentification",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}